pyproject.toml
[tool.poetry]
name = "telegrambot"
version = "0.1.0"
description = ""
authors = ["Zhangeldi <zhanikshakimurat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
httpx = "^0.28.1"
pydantic = "^2.11.3"
openai = "^1.72.0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
